Add "grpc_code" label for grpc_{client,server}_handled_total
The go-grpc-prometheus and python-grpc-prometheus libraries use the name
"grpc_code" to record the GRPC message code of handled calls.

To avoid breaking compatibility with users that depend on the "code"
label being used, we add a copy of the label under the name "grpc_code".
Future releases should deprecate and remove the "code" label.